Add segment register source and test
Segment register has 6 registers inside of it. Write Select picks
which register is written to during writeback. All register values
are outputted. The desired segment register can be muxed later.

Testbench writes a value to each segment regsiter and verifies that
only that segment register is changed. It also tests doing a write
to no register and verifies that no other registers are changed.